# src/docs/openapi2/swagger.yaml
definitions:
...
  dto.BatchDeleteInjectionReq:
    properties:
      ids:
        description: List of injection IDs for deletion
        items:
          type: integer
        type: array
      labels:
        description: List of label keys to match for deletion
        items:
          $ref: '#/definitions/dto.LabelItem'
        type: array
    type: object
...
  dto.GenericResponse-any:
    properties:
      code:
        description: Status code
        type: integer
      data:
        description: Generic type data
      message:
        description: Response message
        type: string
      timestamp:
        description: Response generation time
        type: integer
    type: object
  dto.GenericResponse-array_dto_InjectionNoIssuesResp:
    properties:
      code:
        description: Status code
        type: integer
      data:
        description: Generic type data
        items:
          $ref: '#/definitions/dto.InjectionNoIssuesResp'
        type: array
      message:
        description: Response message
        type: string
      timestamp:
        description: Response generation time
        type: integer
    type: object
...
  dto.GenericResponse-dto_InjectionDetailResp:
    properties:
      code:
        description: Status code
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/dto.InjectionDetailResp'
        description: Generic type data
      message:
        description: Response message
        type: string
      timestamp:
        description: Response generation time
        type: integer
    type: object
...
  dto.GenericResponse-dto_InjectionResp:
    properties:
      code:
        description: Status code
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/dto.InjectionResp'
        description: Generic type data
      message:
        description: Response message
        type: string
      timestamp:
        description: Response generation time
        type: integer
    type: object
...
  dto.GenericResponse-dto_ListResp-dto_InjectionResp:
    properties:
      code:
        description: Status code
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/dto.ListResp-dto_InjectionResp'
        description: Generic type data
      message:
        description: Response message
        type: string
      timestamp:
        description: Response generation time
        type: integer
    type: object
...
  dto.ManageInjectionLabelReq:
    properties:
      add_labels:
        description: List of labels to add
        items:
          $ref: '#/definitions/dto.LabelItem'
        type: array
      remove_labels:
        description: List of label keys to remove
        items:
          type: string
        type: array
    type: object
...
host: localhost:8082
info:
  contact:
    email: team@rcabench.com
    name: RCABench Team
  description: |-
    RCABench - A comprehensive root cause analysis benchmarking platform for microservices
    This API provides endpoints for managing datasets, algorithms, evaluations, and fault injections
    for root cause analysis in distributed systems and microservices architectures.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: RCABench API
  version: 1.0.1
paths:
...
  /api/v2/injections:
    get:
      description: Get a paginated list of injections with pagination and filtering
      operationId: list_injections
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: size
        type: integer
      - description: Filter by fault type
        in: query
        name: fault_type
        type: integer
      - description: Filter by benchmark
        in: query
        name: benchmark
        type: string
      - description: Filter by event
        in: query
        name: event
        type: integer
      - description: Filter by status
        in: query
        name: status
        type: integer
      - collectionFormat: csv
        description: Filter by labels (array of key:value strings, e.g., 'type:chaos')
        in: query
        items:
          type: string
        name: labels
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: Injections retrieved successfully
          schema:
            $ref: '#/definitions/dto.GenericResponse-dto_ListResp-dto_InjectionResp'
        "400":
          description: Invalid request format or parameters
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "403":
          description: Permission denied
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
      security:
      - BearerAuth: []
      summary: List injections
      tags:
      - Injections
  /api/v2/injections/{id}:
    get:
      description: Get detailed information about a specific injection
      operationId: get_injection_by_id
      parameters:
      - description: Injection ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Injection retrieved successfully
          schema:
            $ref: '#/definitions/dto.GenericResponse-dto_InjectionDetailResp'
        "400":
          description: Invalid injection ID
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "403":
          description: Permission denied
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "404":
          description: Injection not found
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
      security:
      - BearerAuth: []
      summary: Get injection by ID
      tags:
      - Injections
  /api/v2/injections/{id}/labels:
    patch:
      consumes:
      - application/json
      description: Add or remove custom labels (key-value pairs) for an injection
      operationId: update_injection_labels
      parameters:
      - description: Injection ID
        in: path
        name: id
        required: true
        type: integer
      - description: Custom label management request
        in: body
        name: manage
        required: true
        schema:
          $ref: '#/definitions/dto.ManageInjectionLabelReq'
      produces:
      - application/json
      responses:
        "200":
          description: Custom labels managed successfully
          schema:
            $ref: '#/definitions/dto.GenericResponse-dto_InjectionResp'
        "400":
          description: Invalid injection ID or request format/parameters
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "403":
          description: Permission denied
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "404":
          description: Injection not found
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
      security:
      - BearerAuth: []
      summary: Manage injection custom labels
      tags:
      - Injections
  /api/v2/injections/analysis/no-issues:
    get:
      description: Query all fault injection records without issues based on time
        range, returning detailed records including configuration information
      operationId: list_successful_injections
      parameters:
      - collectionFormat: csv
        description: Filter by labels (array of key:value strings, e.g., 'type:chaos')
        in: query
        items:
          type: string
        name: labels
        type: array
      - description: Time range query, supports custom relative time (1h/24h/7d) or
          custom, default not set
        in: query
        name: lookback
        type: string
      - description: Custom start time, RFC3339 format, required when lookback=custom
        format: date-time
        in: query
        name: custom_start_time
        type: string
      - description: Custom end time, RFC3339 format, required when lookback=custom
        format: date-time
        in: query
        name: custom_end_time
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully returned fault injection records without issues
          schema:
            $ref: '#/definitions/dto.GenericResponse-array_dto_InjectionNoIssuesResp'
        "400":
          description: Request parameter error, such as incorrect time format or parameter
            validation failure, etc.
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
      summary: Query Fault Injection Records Without Issues
      tags:
      - Injections
  /api/v2/injections/analysis/with-issues:
    get:
      description: Query all fault injection records with issues based on time range
      operationId: list_failed_injections
      parameters:
      - collectionFormat: csv
        description: Filter by labels (array of key:value strings, e.g., 'type:chaos')
        in: query
        items:
          type: string
        name: labels
        type: array
      - description: Time range query, supports custom relative time (1h/24h/7d) or
          custom, default not set
        in: query
        name: lookback
        type: string
      - description: Custom start time, RFC3339 format, required when lookback=custom
        format: date-time
        in: query
        name: custom_start_time
        type: string
      - description: Custom end time, RFC3339 format, required when lookback=custom
        format: date-time
        in: query
        name: custom_end_time
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GenericResponse-array_dto_InjectionWithIssuesResp'
        "400":
          description: Request parameter error, such as incorrect time format or parameter
            validation failure, etc.
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
      summary: Query Fault Injection Records With Issues
      tags:
      - Injections
  /api/v2/injections/batch-delete:
    post:
      consumes:
      - application/json
      description: Batch delete injections by IDs or labels or tags with cascading
        deletion of related records
      operationId: batch_delete_injections
      parameters:
      - description: Batch delete request
        in: body
        name: batch_delete
        required: true
        schema:
          $ref: '#/definitions/dto.BatchDeleteInjectionReq'
      produces:
      - application/json
      responses:
        "200":
          description: Injections deleted successfully
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "403":
          description: Permission denied
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
      security:
      - BearerAuth: []
      summary: Batch delete injections
      tags:
      - Injections
  /api/v2/injections/build:
    post:
      consumes:
      - application/json
      operationId: build_datapack
      parameters:
      - description: Datapack building request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.SubmitDatapackBuildingReq'
      produces:
      - application/json
      responses:
        "202":
          description: Datapack building submitted successfully
          schema:
            $ref: '#/definitions/dto.GenericResponse-dto_SubmitDatapackBuildingResp'
        "400":
          description: Invalid request format or parameters
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "403":
          description: Permission denied
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "404":
          description: Resource not found
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
      summary: Submit batch datapack buildings
      tags:
      - Injections
  /api/v2/injections/inject:
    post:
      consumes:
      - application/json
      description: Submit multiple fault injection tasks in batch
      operationId: inject_fault
      parameters:
      - description: Fault injection request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.SubmitInjectionReq'
      produces:
      - application/json
      responses:
        "200":
          description: Fault injection submitted successfully
          schema:
            $ref: '#/definitions/dto.GenericResponse-dto_SubmitInjectionResp'
        "400":
          description: Invalid request format or parameters
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "403":
          description: Permission denied
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "404":
          description: Resource not found
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
      security:
      - BearerAuth: []
      summary: Submit batch fault injections
      tags:
      - Injections
  /api/v2/injections/metadata:
    get:
      description: Get injection-related metadata including configuration, field mappings,
        and namespace resources
      operationId: get_injection_metadata
      parameters:
      - description: Namespace prefix for config and resources metadata
        in: query
        name: namespace
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully returned metadata
          schema:
            $ref: '#/definitions/dto.GenericResponse-dto_InjectionMetadataResp'
        "400":
          description: Invalid namespace prefix
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "403":
          description: Permission denied
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "404":
          description: Resource not found
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
      summary: Get Injection Metadata
      tags:
      - Injections
  /api/v2/injections/search:
    post:
      consumes:
      - application/json
      description: Advanced search for injections with complex filtering including
        custom labels
      operationId: search_injections
      parameters:
      - description: Search criteria
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/dto.SearchInjectionReq'
      produces:
      - application/json
      responses:
        "200":
          description: Search results
          schema:
            $ref: '#/definitions/dto.GenericResponse-dto_SearchResp-dto_InjectionResp'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "403":
          description: Permission denied
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
      security:
      - BearerAuth: []
      summary: Search injections
      tags:
      - Injections
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"