# src/handlers/v2/injections.go
package v2

import (
	"aegis/consts"
	"context"
	"fmt"
	"net/http"
	"strconv"

	"aegis/dto"
	"aegis/handlers"
	"aegis/middleware"
	producer "aegis/service/prodcuer"
	"aegis/utils"

	"github.com/gin-gonic/gin"
	"github.com/sirupsen/logrus"
	"go.opentelemetry.io/otel/codes"
	"go.opentelemetry.io/otel/trace"

	chaos "github.com/LGU-SE-Internal/chaos-experiment/handler"
)

// BatchDeleteInjections
//
//	@Summary		Batch delete injections
//	@Description	Batch delete injections by IDs or labels or tags with cascading deletion of related records
//	@Tags			Injections
//	@ID				batch_delete_injections
//	@Accept			json
//	@Produce		json
//	@Security		BearerAuth
//	@Param			batch_delete	body		dto.BatchDeleteInjectionReq	true	"Batch delete request"
//	@Success		200				{object}	dto.GenericResponse[any]	"Injections deleted successfully"
//	@Failure		400				{object}	dto.GenericResponse[any]	"Invalid request"
//	@Failure		401				{object}	dto.GenericResponse[any]	"Authentication required"
//	@Failure		403				{object}	dto.GenericResponse[any]	"Permission denied"
//	@Failure		500				{object}	dto.GenericResponse[any]	"Internal server error"
//	@Router			/api/v2/injections/batch-delete [post]
func BatchDeleteInjections(c *gin.Context) {
	var req dto.BatchDeleteInjectionReq
	if err := c.ShouldBindJSON(&req); err != nil {
		dto.ErrorResponse(c, http.StatusBadRequest, "Invalid request: "+err.Error())
		return
	}

	if err := req.Validate(); err != nil {
		dto.ErrorResponse(c, http.StatusBadRequest, "Validation failed: "+err.Error())
		return
	}

	var err error
	if len(req.IDs) > 0 {
		err = producer.BatchDeleteInjectionsByIDs(req.IDs)
	} else {
		err = producer.BatchDeleteInjectionsByLabels(req.Labels)
	}

	if handlers.HandleServiceError(c, err) {
		return
	}

	dto.JSONResponse[any](c, http.StatusNoContent, "Injections deleted successfully", nil)
}

// GetInjection handles getting a single injection by ID
//
//	@Summary		Get injection by ID
//	@Description	Get detailed information about a specific injection
//	@Tags			Injections
//	@ID				get_injection_by_id
//	@Produce		json
//	@Security		BearerAuth
//	@Param			id	path		int												true	"Injection ID"
//	@Success		200	{object}	dto.GenericResponse[dto.InjectionDetailResp]	"Injection retrieved successfully"
//	@Failure		400	{object}	dto.GenericResponse[any]						"Invalid injection ID"
//	@Failure		401	{object}	dto.GenericResponse[any]						"Authentication required"
//	@Failure		403	{object}	dto.GenericResponse[any]						"Permission denied"
//	@Failure		404	{object}	dto.GenericResponse[any]						"Injection not found"
//	@Failure		500	{object}	dto.GenericResponse[any]						"Internal server error"
//	@Router			/api/v2/injections/{id} [get]
func GetInjection(c *gin.Context) {
	idStr := c.Param(consts.URLPathID)
	id, err := strconv.Atoi(idStr)
	if err != nil {
		dto.ErrorResponse(c, http.StatusBadRequest, "Invalid injection ID")
		return
	}

	resp, err := producer.GetInjectionDetail(id)
	if handlers.HandleServiceError(c, err) {
		return
	}

	dto.SuccessResponse(c, resp)
}

// GetInjectionMetadata
//
//	@Summary		Get Injection Metadata
//	@Description	Get injection-related metadata including configuration, field mappings, and namespace resources
//	@Tags			Injections
//	@ID				get_injection_metadata
//	@Produce		json
//	@Param			namespace	query		string											true	"Namespace prefix for config and resources metadata"
//	@Success		200			{object}	dto.GenericResponse[dto.InjectionMetadataResp]	"Successfully returned metadata"
//	@Failure		400			{object}	dto.GenericResponse[any]						"Invalid namespace prefix"
//	@Failure		401			{object}	dto.GenericResponse[any]						"Authentication required"
//	@Failure		403			{object}	dto.GenericResponse[any]						"Permission denied"
//	@Failure		404			{object}	dto.GenericResponse[any]						"Resource not found"
//	@Failure		500			{object}	dto.GenericResponse[any]						"Internal server error"
//	@Router			/api/v2/injections/metadata [get]
func GetInjectionMetadata(c *gin.Context) {
	nsPrefix := c.Query("namespace")
	if !utils.CheckNsPrefixExists(nsPrefix) {
		dto.ErrorResponse(c, http.StatusBadRequest, "Invalid namespace prefix")
		return
	}

	confNode, err := chaos.StructToNode[chaos.InjectionConf](nsPrefix)
	if err != nil {
		handlers.HandleServiceError(c, err)
		return
	}

	faultResourceMap, err := chaos.GetChaosResourceMap()
	if err != nil {
		handlers.HandleServiceError(c, err)
		return
	}

	resourcesMap, err := chaos.GetNsResources()
	if err != nil {
		handlers.HandleServiceError(c, err)
		return
	}

	resources, exists := resourcesMap[nsPrefix]
	if !exists {
		dto.ErrorResponse(c, http.StatusNotFound, "Namespace resources not found")
		return
	}

	dto.SuccessResponse(c, &dto.InjectionMetadataResp{
		Config:           confNode,
		FaultTypeMap:     chaos.ChaosTypeMap,
		FaultResourceMap: faultResourceMap,
		NsResources:      resources,
	})
}

// ListInjections handles listing injections with pagination and filtering
//
//	@Summary		List injections
//	@Description	Get a paginated list of injections with pagination and filtering
//	@Tags			Injections
//	@ID				list_injections
//	@Produce		json
//	@Security		BearerAuth
//	@Param			page		query		int														false	"Page number"	default(1)
//	@Param			size		query		int														false	"Page size"		default(20)
//	@Param			fault_type	query		int														false	"Filter by fault type"
//	@Param			benchmark	query		string													false	"Filter by benchmark"
//	@Param			event		query		int														false	"Filter by event"
//	@Param			status		query		int														false	"Filter by status"
//	@Param			labels		query		[]string												false	"Filter by labels (array of key:value strings, e.g., 'type:chaos')"
//	@Success		200			{object}	dto.GenericResponse[dto.ListResp[dto.InjectionResp]]	"Injections retrieved successfully"
//	@Failure		400			{object}	dto.GenericResponse[any]								"Invalid request format or parameters"
//	@Failure		401			{object}	dto.GenericResponse[any]								"Authentication required"
//	@Failure		403			{object}	dto.GenericResponse[any]								"Permission denied"
//	@Failure		500			{object}	dto.GenericResponse[any]								"Internal server error"
//	@Router			/api/v2/injections [get]
func ListInjections(c *gin.Context) {
	var req dto.ListInjectionReq
	if err := c.ShouldBindQuery(&req); err != nil {
		dto.ErrorResponse(c, http.StatusBadRequest, "Invalid request format: "+err.Error())
		return
	}

	if err := req.Validate(); err != nil {
		dto.ErrorResponse(c, http.StatusBadRequest, "Invalid request parameters: "+err.Error())
		return
	}

	resp, err := producer.ListInjections(&req)
	if handlers.HandleServiceError(c, err) {
		return
	}

	dto.SuccessResponse(c, resp)
}

// TODO: implement the actual search logic
// SearchInjection
//
//	@Summary		Search injections
//	@Description	Advanced search for injections with complex filtering including custom labels
//	@Tags			Injections
//	@ID				search_injections
//	@Accept			json
//	@Produce		json
//	@Security		BearerAuth
//	@Param			search	body		dto.SearchInjectionReq									true	"Search criteria"
//	@Success		200		{object}	dto.GenericResponse[dto.SearchResp[dto.InjectionResp]]	"Search results"
//	@Failure		400		{object}	dto.GenericResponse[any]								"Invalid request"
//	@Failure		401		{object}	dto.GenericResponse[any]								"Authentication required"
//	@Failure		403		{object}	dto.GenericResponse[any]								"Permission denied"
//	@Failure		500		{object}	dto.GenericResponse[any]								"Internal server error"
//	@Router			/api/v2/injections/search [post]
func SearchInjections(c *gin.Context) {
	var req dto.SearchInjectionReq
	if err := c.ShouldBindJSON(&req); err != nil {
		dto.ErrorResponse(c, http.StatusBadRequest, "Invalid request: "+err.Error())
		return
	}

	if err := req.Validate(); err != nil {
		dto.ErrorResponse(c, http.StatusBadRequest, "Validation failed: "+err.Error())
		return
	}
}

// ManageInjectionCustomLabels manages injection custom labels (key-value pairs)
//
//	@Summary		Manage injection custom labels
//	@Description	Add or remove custom labels (key-value pairs) for an injection
//	@Tags			Injections
//	@ID				update_injection_labels
//	@Accept			json
//	@Produce		json
//	@Security		BearerAuth
//	@Param			id		path		int										true	"Injection ID"
//	@Param			manage	body		dto.ManageInjectionLabelReq				true	"Custom label management request"
//	@Success		200		{object}	dto.GenericResponse[dto.InjectionResp]	"Custom labels managed successfully"
//	@Failure		400		{object}	dto.GenericResponse[any]				"Invalid injection ID or request format/parameters"
//	@Failure		401		{object}	dto.GenericResponse[any]				"Authentication required"
//	@Failure		403		{object}	dto.GenericResponse[any]				"Permission denied"
//	@Failure		404		{object}	dto.GenericResponse[any]				"Injection not found"
//	@Failure		500		{object}	dto.GenericResponse[any]				"Internal server error"
//	@Router			/api/v2/injections/{id}/labels [patch]
func ManageInjectionCustomLabels(c *gin.Context) {
	idStr := c.Param(consts.URLPathID)
	id, err := strconv.Atoi(idStr)
	if err != nil {
		dto.ErrorResponse(c, http.StatusBadRequest, "Invalid injection ID")
		return
	}

	var req dto.ManageInjectionLabelReq
	if err := c.ShouldBindJSON(&req); err != nil {
		dto.ErrorResponse(c, http.StatusBadRequest, "Invalid request format: "+err.Error())
		return
	}

	if err := req.Validate(); err != nil {
		dto.ErrorResponse(c, http.StatusBadRequest, "Invalid request parameters: "+err.Error())
		return
	}

	resp, err := producer.ManageInjectionLabels(&req, id)
	if handlers.HandleServiceError(c, err) {
		return
	}

	dto.SuccessResponse(c, resp)
}

// SubmitFaultInjection submits batch fault injections
//
//	@Summary		Submit batch fault injections
//	@Description	Submit multiple fault injection tasks in batch
//	@Tags			Injections
//	@ID				inject_fault
//	@Accept			json
//	@Produce		json
//	@Security		BearerAuth
//	@Param			body	body		dto.SubmitInjectionReq							true	"Fault injection request body"
//	@Success		200		{object}	dto.GenericResponse[dto.SubmitInjectionResp]	"Fault injection submitted successfully"
//	@Failure		400		{object}	dto.GenericResponse[any]						"Invalid request format or parameters"
//	@Failure		401		{object}	dto.GenericResponse[any]						"Authentication required"
//	@Failure		403		{object}	dto.GenericResponse[any]						"Permission denied"
//	@Failure		404		{object}	dto.GenericResponse[any]						"Resource not found"
//	@Failure		500		{object}	dto.GenericResponse[any]						"Internal server error"
//	@Router			/api/v2/injections/inject [post]
func SubmitFaultInjection(c *gin.Context) {
	groupID := c.GetString("groupID")
	userID, exists := middleware.GetCurrentUserID(c)
	if !exists {
		dto.ErrorResponse(c, http.StatusUnauthorized, "Authentication required")
		return
	}

	ctx, ok := c.Get(middleware.SpanContextKey)
	if !ok {
		dto.ErrorResponse(c, http.StatusInternalServerError, "Internal server error")
		return
	}

	spanCtx := ctx.(context.Context)
	span := trace.SpanFromContext(spanCtx)

	var req dto.SubmitInjectionReq
	if err := c.BindJSON(&req); err != nil {
		span.SetStatus(codes.Error, "validation error in SubmitFaultInjection: "+err.Error())
		dto.ErrorResponse(c, http.StatusBadRequest, "Invalid request format: "+err.Error())
		return
	}

	if err := req.Validate(); err != nil {
		span.SetStatus(codes.Error, "validation error in SubmitFaultInjection: "+err.Error())
		dto.ErrorResponse(c, http.StatusBadRequest, "Invalid request parameters: "+err.Error())
		return
	}

	resp, err := producer.ProduceRestartPedestalTasks(spanCtx, &req, groupID, userID)
	if handlers.HandleServiceError(c, err) {
		span.SetStatus(codes.Error, "panic in SubmitFaultInjection")
		return
	}

	span.SetStatus(codes.Ok, fmt.Sprintf("Successfully submitted %d fault injections with groupID: %s", len(resp.Items), groupID))
	dto.SuccessResponse(c, resp)
}

// SubmitDatapackBuilding submits batch datapack buildings
//
//	@Summary		Submit batch datapack buildings
//	@Description.	Submit multiple datapack building tasks in batch
//	@Tags			Injections
//	@ID				build_datapack
//	@Accept			json
//	@Produce		json
//	@Param			body	body		dto.SubmitDatapackBuildingReq						true	"Datapack building request body"
//	@Success		202		{object}	dto.GenericResponse[dto.SubmitDatapackBuildingResp]	"Datapack building submitted successfully"
//	@Failure		400		{object}	dto.GenericResponse[any]							"Invalid request format or parameters"
//	@Failure		401		{object}	dto.GenericResponse[any]							"Authentication required"
//	@Failure		403		{object}	dto.GenericResponse[any]							"Permission denied"
//	@Failure		404		{object}	dto.GenericResponse[any]							"Resource not found"
//	@Failure		500		{object}	dto.GenericResponse[any]							"Internal server error"
//	@Router			/api/v2/injections/build [post]
func SubmitDatapackBuilding(c *gin.Context) {
	groupID := c.GetString("groupID")
	userID, exists := middleware.GetCurrentUserID(c)
	if !exists {
		dto.ErrorResponse(c, http.StatusUnauthorized, "Authentication required")
		return
	}

	ctx, ok := c.Get(middleware.SpanContextKey)
	if !ok {
		dto.ErrorResponse(c, http.StatusInternalServerError, "Internal server error")
		return
	}

	spanCtx := ctx.(context.Context)
	span := trace.SpanFromContext(spanCtx)

	var req dto.SubmitDatapackBuildingReq
	if err := c.BindJSON(&req); err != nil {
		span.SetStatus(codes.Error, "validation error in SubmitDatapackBuilding: "+err.Error())
		dto.ErrorResponse(c, http.StatusBadRequest, "Invalid request format: "+err.Error())
		return
	}

	if err := req.Validate(); err != nil {
		span.SetStatus(codes.Error, "validation error in SubmitDatapackBuilding: "+err.Error())
		dto.ErrorResponse(c, http.StatusBadRequest, "Invalid request parameters: "+err.Error())
		return
	}

	resp, err := producer.ProduceDatapackBuildingTasks(spanCtx, &req, groupID, userID)
	if handlers.HandleServiceError(c, err) {
		return
	}

	span.SetStatus(codes.Ok, fmt.Sprintf("Successfully submitted %d datapack buildings with groupID: %s", len(resp.Items), groupID))
	dto.SuccessResponse(c, resp)
}

// ListFaultInjectionNoIssues
//
//	@Summary		Query Fault Injection Records Without Issues
//	@Description	Query all fault injection records without issues based on time range, returning detailed records including configuration information
//	@Tags			Injections
//	@ID				list_successful_injections
//	@Produce		json
//	@Param			labels				query		[]string											false	"Filter by labels (array of key:value strings, e.g., 'type:chaos')"
//	@Param			lookback			query		string												false	"Time range query, supports custom relative time (1h/24h/7d) or custom, default not set"
//	@Param			custom_start_time	query		string												false	"Custom start time, RFC3339 format, required when lookback=custom"	Format(date-time)
//	@Param			custom_end_time		query		string												false	"Custom end time, RFC3339 format, required when lookback=custom"	Format(date-time)
//	@Success		200					{object}	dto.GenericResponse[[]dto.InjectionNoIssuesResp]	"Successfully returned fault injection records without issues"
//	@Failure		400					{object}	dto.GenericResponse[any]							"Request parameter error, such as incorrect time format or parameter validation failure, etc."
//	@Failure		500					{object}	dto.GenericResponse[any]							"Internal server error"
//	@Router			/api/v2/injections/analysis/no-issues [get]
func ListFaultInjectionNoIssues(c *gin.Context) {
	var req dto.ListInjectionNoIssuesReq
	if err := c.BindQuery(&req); err != nil {
		logrus.Errorf("failed to bind query parameters: %v", err)
		dto.ErrorResponse(c, http.StatusBadRequest, "Invalid query parameters")
		return
	}

	if err := req.Validate(); err != nil {
		logrus.Errorf("invalid query parameters: %v", err)
		dto.ErrorResponse(c, http.StatusBadRequest, err.Error())
		return
	}

	items, err := producer.ListInjectionsNoIssues(&req)
	if handlers.HandleServiceError(c, err) {
		return
	}

	dto.SuccessResponse(c, items)
}

// ListFaultInjectionWithIssues
//
//	@Summary		Query Fault Injection Records With Issues
//	@Description	Query all fault injection records with issues based on time range
//	@Tags			Injections
//	@ID				list_failed_injections
//	@Produce		json
//	@Param			labels				query		[]string	false	"Filter by labels (array of key:value strings, e.g., 'type:chaos')"
//	@Param			lookback			query		string		false	"Time range query, supports custom relative time (1h/24h/7d) or custom, default not set"
//	@Param			custom_start_time	query		string		false	"Custom start time, RFC3339 format, required when lookback=custom"	Format(date-time)
//	@Param			custom_end_time		query		string		false	"Custom end time, RFC3339 format, required when lookback=custom"	Format(date-time)
//	@Success		200					{object}	dto.GenericResponse[[]dto.InjectionWithIssuesResp]
//	@Failure		400					{object}	dto.GenericResponse[any]	"Request parameter error, such as incorrect time format or parameter validation failure, etc."
//	@Failure		500					{object}	dto.GenericResponse[any]	"Internal server error"
//	@Router			/api/v2/injections/analysis/with-issues [get]
func ListFaultInjectionWithIssues(c *gin.Context) {
	var req dto.ListInjectionWithIssuesReq
	if err := c.BindQuery(&req); err != nil {
		logrus.Errorf("failed to bind query parameters: %v", err)
		dto.ErrorResponse(c, http.StatusBadRequest, "Invalid query parameters")
		return
	}

	if err := req.Validate(); err != nil {
		logrus.Errorf("invalid query parameters: %v", err)
		dto.ErrorResponse(c, http.StatusBadRequest, err.Error())
		return
	}

	items, err := producer.ListInjectionsWithIssues(&req)
	if handlers.HandleServiceError(c, err) {
		return
	}

	dto.SuccessResponse(c, items)
}